feat: centralise version validation and redirects
We will be adding more endpoints to this api soon, and the current
configuration required a lot of different routes to handle redirecting
to the latest version. This patch bundles that validation and
redirecting logic into a semi-route-agnostic decorator that should be
applicable for most of the endpoints we need to create. It expects any
wrapped view to provide two view args: `dataset_id` and `version`.