build(deps): upgrade dependencies
... that have reported vulnrabilities

The changes in the tests, closing http connections, are not for completely
known reasons - without this the tests time out waiting the server under test
to exit. It's supspected that without this Python requests keeps connections
open, which in turn makes the Gevent web server keep its greenlets running,
that makes them wait forever when attempting to exit the sever gracefully.

Why this happened now and not before, it's not known.

However, this change should be a fairly safe thing to do because in GOV.UK PaaS
the last hop of the HTTP chains to the server through the various layers are
_not_ kept alive - a new TCP connection is made on each HTTP request. Thus the
tests are now more like the PaaS environment than they were before.