feat: multi-statement reports expose data from each statement
The Public Data API has the concept of "reports" in SQLite-based datasets.
These are SQL scripts that define some sort of filtered or aggregated "report"
on the data in the SQLite file where each report's data is exposed as a single
ODS file, and and single CSV file.

The SQL scripts can be made of multiple statements, for example to create
temporary table(s). Then the final statement is a SELECT to query over the
temporary table(s) and/or the real tables.

The code, before this change, dealt with multiple SELECT statements by only
ever exposing the data in the final SELECT statement as the data in the report.

This is now changed. Now if there are multiple SELECT statements in a report
(that result in rows), they are all output:

- For the CSV output, all the rows from all the SELECTs are concatenated
  together
- For ODS output, each SELECT's rows are output in another sheet.

While this is a technically a breaking change in the code, the instance of the
Public Data API at https://data.api.trade.gov.uk/ doesn't as yet have any
reports that have multiple SELECT statements, so until they do, this won't
introduce any different behaviour.

The use case for this is the UK Tariff, usually accessed via
https://www.data.gov.uk/dataset/3bee9a8a-e69c-400e-add5-3345a87a8e25/tariffs-to-trade-with-the-uk-from-1-january-2021.
At the moment each of the 21 "Sections" of the tariff's "Measures on Declarable
Commodities" report" is a separate report, which means to download the whole
thing the user has to download 21 files which is not convenient.  This change
allows all 21 sections to be in a single file. They are not all in the one ODS
sheet, because this makes it not usable in Excel - all sections in one sheet go
beyond Excel's limit on how many rows are in one sheet. When split into 21
sheets in a single ODS file, this makes it usable in Excel.

The single CSV file is most likely not openable in Excel, but given the ODS
file should be, this is probably acceptable and a step forward because
non-Excel CSV-users should be able to open the single file.