feat: add monitoring thread to worker that writes to heartbeat file
This makes the worker write the current timestamp (seconds since epoch) to a
file. This isn't used right now, but will be used when we move to DBT Platform.

There might be a follow up change - a command line to run to check if this file
exists and has a recent timestamp.