build(deps): move to Python 3.11
In preparation to moving to DBT platform, upgarding dependencies. And in this
case, moving Python from 3.9 to 3.11.

Opting, for now, to not move to 3.12 to avoid having the latest and greatest to
minimise risk of bugs.

The upgrade of pylint and astroid, and ElasticAPM is to avoid the error at
various points of:

> ImportError: cannot import name 'formatargspec' from 'inspect'
> (/home/circleci/.pyenv/versions/3.11.9/lib/python3.11/inspect.py)

when running pylint. And the pylint configuration changes, and small code
changes are as a consequence of the new version of pylint having slightly
different default rules.

The upgrade of coverage, numpy and pandas is to avoid the error

> E   ImportError: numpy.core.multiarray failed to import

when running tests with code coverage in CI