feat: better worker logging
The logging from the worker was logging on every http request, and because
querying from SQLite files uses _many_ requests, the logs were too noisy to be
helpful.

To address this, the httpx log level in the worker is increased from INFO to
WARNING, and manual log messages have been put into key parts of the process.