feat: add GA4 server-side analytics
This adds GA4 server-side analytics using the GA4 measurement protocol, similar
to the UA measurement protocol already in place.

Both will be run at the same time for now as GA4 is tested - and there may well
be more changes to the GA4 server-side analytics in the next week or so.

Note that server-side is fairly important because the primary function of the
public-data-api is the API (which includes data download), which do not have
web pages in front of them to run client-side analytics.

Co-authored-by: Ian Leggett <contactleggett@gmail.com>
Co-authored-by: Michal Charemza <michal@charemza.name>