fix: march forward with conversion if cannot convert to Parquet
This change wraps the code that converts to Parquet format with a try block to
march forward in the case of ArrowNotImplementedError. This is to work around
the error:

> ArrowNotImplementedError ("NumPyConverter doesn't implement <null> conversion."
> , 'Conversion failed for column <column_name> with type float64')

(The "<column_name>" in the real error has the name of the column).

That I suspect is due to Pandas schema detection assuming a column only has
null in because it only null appears at the beginning, but then non-null values
appear later.

This is causing a fair bit of noise in logs. Since we only introduced Parquet
recently, for now I think this is the best option because it's not taking
Parquet away from any files that already have it.